eatgo-api/src/main/java/kr/co/fastcampus/eatgo/interfaces/ReviewController.java
package kr.co.fastcampus.eatgo.interfaces;

import kr.co.fastcampus.eatgo.application.ReviewService;
import kr.co.fastcampus.eatgo.domain.Review;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.ResponseBody;

import java.net.URI;
import java.net.URISyntaxException;

@Controller
public class ReviewController {
    @Autowired
    private ReviewService reviewService;

    @PostMapping("/restaurants/{restaurantId}/reviews")
    public ResponseEntity<?> create(
            @PathVariable("restaurantId") Long id,
            @RequestBody Review resource
    ) throws URISyntaxException {
        Review review = reviewService.addReview(id, resource);

        String url = "/restaurants/"+id+
                "/reviews/"+ review.getId();

        return ResponseEntity.created(new URI(url))
                .body("{}");
    }
}
